* master:
  fuse_lowlevel.c: drop incorrect comment from convert_to_conn_want_ext()
  fuse_lowlevel.c: fix possible 64 bits value truncation
  github actions: Add an include-what-you-need check
  build(deps): bump github/codeql-action from 3.28.1 to 3.28.6
  lib: Set thread names
  abidiff: Don't warn about added symbols
  checkpatch: Ignore ENOSYS_SYSCALL
  codespell: Add 're-used' and 're-using' to the ignore list
  build(deps): bump github/codeql-action from 3.28.0 to 3.28.1
  Update doxygen/comments for fuse_reply_open/fuse_reply_create
  Identify the FD hold by parent process

+++ lib/fuse_lowlevel.c
@@ -2393,8 +2393,6 @@ static bool want_flags_valid(uint64_t capable, uint64_t want)
 
 /**
  * Get the wanted capability flags, converting from old format if necessary
- * Also applies the first 32 bits of capable_ext to capable
- *
  */
 static inline int convert_to_conn_want_ext(struct fuse_conn_info *conn,
 					   uint64_t want_ext_default)
@@ -2577,7 +2575,7 @@ _do_init(fuse_req_t req, const fuse_ino_t nodeid, const void *op_in,
 
 	se->got_init = 1;
 	if (se->op.init) {
-		uint32_t want_ext_default = se->conn.want_ext;
+		uint64_t want_ext_default = se->conn.want_ext;
 		int rc;
 
 		// Apply the first 32 bits of capable_ext to capable
